Add files via upload
This is the application that I added to a larger application for my Prosper IT internship. It is a dB that allows the user to add their favorite song, view it and other entries, and edit or delete them. There is an API added for the current temperature of an area in London.
It was created using the Django Framework.